Pull transaction table out the pull info container
